Move loader and server to ./src
